alloc: Only reserve 1 UL slot with algorithm B (TODO)
Since currently the algorithm B will only allocate a single UL slot
and will have to stick to it (first common TS), the other possible UL
slots will not be allocated while the reservation is kept.

This commit adds code to update the reserved set of UL slots to the
single common TS when the UL TBF is allocated.

TODO:
- test it

Sponsored-by: On-Waves ehf