edge: Get EGPRS MS class from downlink BSSGP (TODO)
In case there is no MS object present that matches a DL UNITDATA,
the EGPRS class is currently assumed to be 0, since the Radio
Access Capabilities IE is not fully decoded if present.

This commit uses the CSN.1 decoder to parse the IE and uses the
same functions like the uplink related code does to get the GPRS
and EGPRS multislot classes.

TODO:
 - test and remove assert
 - remove old function

Sponsored-by: On-Waves ehf